Rename grid modules to data_grid for clarity
Renamed L0_granule_grid.py to L0_granule_data_grid.py and L1_field_grid.py to L1_field_data_grid.py for improved naming consistency. Updated all relevant imports and adjusted the spacetime package docstring to reflect these changes.
