+++ openwave/spacetime/__init__.py
@@ -2,6 +2,6 @@
 Energy Source modules for OpenWave simulations.
 
 This package contains the fundamental components of Spacetime simulation:
-- data_grid: Wave-Field representations and calculations
-- energy_wave: Energy-Wave representations and calculations
+- data_grids: Wave data structures & representations
+- wave_engines: Energy-Wave representations and calculations
 """
