+++ openwave/validations/stability_analysis_wave.py
@@ -7,7 +7,7 @@ import numpy as np
 
 from openwave.common import constants
 
-import openwave.spacetime.L1_field_grid as data_grid
+import openwave.spacetime.L1_field_data_grid as data_grid
 
 ti.init(arch=ti.gpu)
