+++ openwave/xperiments/5_L1_field_based/_L1_launcher_archive.py
@@ -19,7 +19,7 @@ import taichi as ti
 from openwave.common import colormap, constants
 from openwave._io import flux_mesh, render, video
 
-import openwave.spacetime.L1_field_grid as data_grid
+import openwave.spacetime.L1_field_data_grid as data_grid
 import openwave.spacetime.L1_wave_engine as ewave
 import _L1_instrumentation as instrument
